requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
BASC-py4chan==0.6.5
certifi==2020.6.20
chardet==3.0.4
discord.py==1.4.1
idna==2.10
multidict==4.7.6
py-4chan==0.2.8
requests==2.24.0
urllib3==1.25.10
yarl==1.5.1
